refactor: integrate util files related to JWT and add more detailed description to utils files
