+++ src/utils.rs
@@ -1,5 +1,4 @@
 pub mod token;
-pub mod decode_jwt;
-pub mod encode_jwt;
+pub mod jwt;
 pub mod oauth_authentication;
 pub mod verify_password;
