+++ src/utils/oauth_authentication.rs
@@ -1,9 +1,17 @@
-// TODO: OAuth処理の作成
-
 use std::{collections::HashMap, time::Duration, env};
 use serde::{Serialize, Deserialize};
 use dotenv::dotenv;
 
+/// ### RefreshResult
+/// アクセストークンを取得するリクエストのレスポンスです
+///
+/// access_token    : アクセストークン
+///
+/// expires_in      : トークンの使用期限（秒）
+///
+/// scope           : access_tokenに寄って付与されるスコープ
+///
+/// token_type      : 常にBearerが指定される
 #[derive(Clone, PartialEq, Serialize, Deserialize)]
 pub struct RefreshResult {
     pub access_token: String,
@@ -12,6 +20,14 @@ pub struct RefreshResult {
     pub token_type: String,
 }
 
+/// ### refresh_access_token
+/// リフレッシュトークンを使用して、Google Oauthのアクセストークンを取得します.
+///
+/// refresh_token   : リフレッシュトークン
+///
+/// client_id       : Oauth Client Id
+///
+/// client_secret   : Oauth Client Secret
 pub async fn refresh_access_token(refresh_token: &str, client_id: &str, client_secret: &str) -> Result<RefreshResult, Box<dyn std::error::Error + Send + Sync>> {
     let mut params = HashMap::new();
